fix(rtf): memoryLimitInKb*1024 int-overflow in the \objdata cap (Gemini #1) + drop swept-in script
- The cumulative \objdata cap compared os.size() against `memoryLimitInKb * 1024`, an
  int*int product that overflows to negative for a large memoryLimitInKb (>= ~2M KB),
  making the check `os.size() >= <negative>` always true (or bypassed). Do the multiply
  in long: `(long) memoryLimitInKb * 1024` — at all sites incl. the pre-existing
  writeBytes cap. TikaMemoryLimitException already takes (long, long).
- Remove generate_qr_ics.py — a local helper a `git add -A` swept into the parser-fix
  commit; untracked (kept on disk) + gitignored.

+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-microsoft-module/src/main/java/org/apache/tika/parser/microsoft/rtf/RTFEmbObjHandler.java
@@ -178,8 +178,8 @@ class RTFEmbObjHandler {
         if (state != EMB_STATE.OBJDATA) {
             return;
         }
-        if (os.size() >= memoryLimitInKb * 1024) {
-            throw new TikaMemoryLimitException(os.size() + 1, memoryLimitInKb * 1024);
+        if (os.size() >= (long) memoryLimitInKb * 1024) {
+            throw new TikaMemoryLimitException(os.size() + 1, (long) memoryLimitInKb * 1024);
         }
         os.write(b);
         hexEscapeInObjdata = true;
@@ -207,8 +207,8 @@ class RTFEmbObjHandler {
                     throw new IOException("hex char to byte overflow");
                 }
 
-                if (os.size() >= memoryLimitInKb * 1024) {
-                    throw new TikaMemoryLimitException(os.size() + 1, memoryLimitInKb * 1024);
+                if (os.size() >= (long) memoryLimitInKb * 1024) {
+                    throw new TikaMemoryLimitException(os.size() + 1, (long) memoryLimitInKb * 1024);
                 }
                 os.write((int) sum);
 
@@ -225,8 +225,8 @@ class RTFEmbObjHandler {
         if (len < 0) {
             throw new TikaException("Requesting I read < 0 bytes ?!");
         }
-        if (len > memoryLimitInKb * 1024) {
-            throw new TikaMemoryLimitException(len, (memoryLimitInKb * 1024));
+        if (len > (long) memoryLimitInKb * 1024) {
+            throw new TikaMemoryLimitException(len, ((long) memoryLimitInKb * 1024));
         }
 
         byte[] bytes = new byte[len];
